fixes for demo, jamfile

+++ Demos/GimpactTestDemo/GimpactTestDemo.cpp
@@ -325,7 +325,7 @@ void	GimpactConcaveDemo::initPhysics()
 	//btOverlappingPairCache* broadphase = new btSimpleBroadphase();
 	//m_broadphase = new btSimpleBroadphase();
 
-	LONG maxProxies = 1024;
+	int  maxProxies = 1024;
 	btVector3 worldAabbMin(-10000,-10000,-10000);
 	btVector3 worldAabbMax( 10000, 10000, 10000);
 	m_broadphase = new bt32BitAxisSweep3(worldAabbMin,worldAabbMax,maxProxies);
